requirements.txt
numpy
pandas
pillow
opencv-python
openai>=1.0.0
python-dotenv
pyyaml
